data-processing/spark/examples/001-scd2-w-delta/pyproject.toml
[project]
name = "001-scd2-w-delta"
version = "0.1.0"
description = "Spark-related tutorials - SCD2 with Delta Lake"
readme = "README.md"
requires-python = ">=3.12,<3.13"
authors = [
    {name = "Denis Arnaud", email = "denis.arnaud_fedora@m4x.org"},
]
urls = {repository = "https://github.com/data-engineering-helpers/ks-cheat-sheets"}

# DEPENDENCIES
dependencies = [
    "datanomy",
    "loguru",
    "PyYAML"
]

[dependency-groups]
dev = [
  "packaging",
  "pyspark[connect,sql,pipelines,pandas_on_spark]",
  "delta-spark",
  "ruff",
  "setuptools",
  "toml",
  "twine",
]
lint = [
  "ruff"
]

[project.scripts]
merge_customer_001_simple = "001_scd2_w_delta.jobs.merge_customer_001_simple:main"

[build-system]
requires = ["uv_build<0.12"]
build-backend = "uv_build"

[tool.ruff]
#src = ["src"]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "Q", "I"]
ignore = ["E203", "E501", "E712"]

[tool.ruff.lint.pydocstyle]
convention = "google"

